requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
streamlit>=1.35.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
beautifulsoup4>=4.12.0
fastapi>=0.110.0
uvicorn>=0.27.0
